Lab9-MCPSnowflake/MealPlanner/requirements.txt
streamlit==1.39.0
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
snowflake-connector-python==3.12.0
requests==2.31.0
python-dotenv==1.0.0